chore: rename View -> CodeView
