+++ crates/knap_edit/src/lib.rs
@@ -1,7 +1,7 @@
+mod code_view;
 mod command_bar;
 mod commands;
 pub mod editor;
 mod highlighter;
 mod message_bar;
 mod status_bar;
-mod view;
